implement the apis to update and delete opportunites

+++ src/models/review.js
@@ -2,7 +2,7 @@ const { DataTypes } = require("sequelize");
 const { sequelize } = require("../config/db");
 const Review = sequelize.define("Review", {
   rating_count: { type: DataTypes.FLOAT, allowNull: false },
-  date: { type: DataTypes.INTEGER, allowNull: false },
+  date: { type: DataTypes.BIGINT, allowNull: false },
   role: { type: DataTypes.STRING },
   review: { type: DataTypes.TEXT },
   profile_picture: { type: DataTypes.STRING },
